fix: multiple credential providers with different ids
